% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-interp-contour.R
\name{geom_interp_contour}
\alias{geom_interp_contour}
\title{Create a contour geometry from irregularly spaced data}
\usage{
geom_interp_contour(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{arguments passed to base function,}

\item{linear}{(boolean) if TRUE, use linear interpolation. Else use cubic (see ?akima::interp)}

\item{duplicate}{(char) how to deal with duplicates (see ?akima::interp)}

\item{dupfun}{(function) if duplicate = "user" (see ?akima::interp)}

\item{rescale}{(boolean) if TRUE, rescale X and Y to the 0-1 range (helps in getting a smooth surface)}
}
\description{
This stat is based on [geom_contour()],
and has the same aesthetics, but is able to operate on irregular data
}
\details{
The function does not like when the range of X and Y data is too dissimilar...
}
