% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-interp-contour.R
\name{stat_interp_contour}
\alias{stat_interp_contour}
\title{Create a contour geometry from irregularly spaced data}
\usage{
stat_interp_contour(
  mapping = NULL,
  data = NULL,
  geom = "contour",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  bins = NULL,
  binwidth = NULL,
  breaks = NULL,
  ...
)
}
\arguments{
\item{...}{arguments passed to base function,}

\item{linear}{(boolean) if TRUE, use linear interpolation. Else use cubic (see ?akima::interp)}
}
\description{
This stat is based on [geom_contour()],
and has the same aesthetics, but is able to operate on irregular data
}
\details{
The function does not like when the range of X and Y data is too dissimilar...
}
