% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-interp-raster.R
\name{stat_interp_raster}
\alias{stat_interp_raster}
\title{Create a raster geometry from irregularly spaced data}
\usage{
stat_interp_raster(
  mapping = NULL,
  data = NULL,
  geom = "raster",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{arguments passed to base function,}
}
\description{
This stat is based on [geom_raster()],
and has the same aesthetics, but is able to operate on irregular data
}
\details{
The function does not like when the range of X and Y data is too dissimilar...
}
